requirements.txt
flask==3.0.0
flask-cors==4.0.0
bcrypt==4.1.2
chromadb==0.4.22
sentence-transformers==2.3.1
requests==2.31.0
werkzeug==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
schedule==1.2.1
